+++ src/app/src/androidTest/java/ca/cmput301f21t22/nabu/documents/IntegrityTests.java
@@ -1,6 +1,7 @@
-package ca.cmput301f21t22.nabu;
+package ca.cmput301f21t22.nabu.documents;
 
 import static org.awaitility.Awaitility.await;
+import static org.junit.Assert.fail;
 
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
@@ -14,7 +15,7 @@ public abstract class IntegrityTests {
 
     @Before
     public void Setup() {
-        this.ref.delete().addOnCompleteListener(task -> this.ready = true);
+        this.ref.delete().addOnCompleteListener(task -> this.ready = true).addOnFailureListener(task -> fail());
         await().until(() -> this.ready);
     }
 }
